@@ -4,13 +4,13 @@ version: 1.7.2
 homepage: https://github.com/smartnsoft/flappy_search_bar
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.16.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   async: ^2.2.0
-  flutter_staggered_grid_view: ^0.3.0
+  flutter_staggered_grid_view: ^0.6.1
 
 dev_dependencies:
   flutter_test:
